refactor: Clean up E2E tests
