+++ ui_automation_tests/step_defs/goods_test.py
@@ -40,10 +40,3 @@ def delete_my_good_in_list(driver):
 def good_is_no_longer_in_list(driver):
     assert context.edited_description not in driver.find_element_by_tag_name("body").text
 
-
-@then('Rollback goods')
-def rollback_goods(driver):
-    if context.good_description in driver.find_element_by_css_selector(".govuk-table__cell").text or context.edited_description in driver.find_element_by_css_selector(".govuk-table__cell").text:
-        goods_list = GoodsList(driver)
-        driver.find_element_by_css_selector("[href*='goods/delete/']").click()
-        goods_list.click_on_delete_good_button()
